+++ src/sim_mirror/core/screen_relay.py
@@ -265,6 +265,8 @@ class ScreenRelay:
                     self._manager.simctl(instance),
                     instance.udid,
                     on_touch=lambda: self._manager.person_touched(instance),
+                    typing=self._config.device_typing,
+                    keyboard_is_us=self._manager.keyboard_is_us,
                 )
                 self._person_session = session
             try: